模組1練習題(新)
模組1練習題(新)

+++ src/p1.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class p1 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in); //使用者輸入(背起來!)
-        int a = sc.nextInt(); //
+        float a = sc.nextFloat(); //
         double b = (double) a ;
         System.out.print(b*1.6);
     }
